refactor(Core/WorldSession): move LoginQueryHolder declaration into the header
